se separa la fecha de la hora en las reservas. El precio se diferencia por tipo de instalacion. Empiezo con el crud de reservas para editar y borrar.

+++ src/main/java/com/example/instalacionesDeportivas/entities/instalaciones.java
@@ -26,7 +26,7 @@ public class instalaciones {
     public String Nombre;
     private String Descripcion;
     private String Imagen;
-    private Float Precio;
+    private double Precio;
     
     @OneToMany(mappedBy = "instalaciones") 
     private List<reservas> reservas;
